Get weather app running in a container with Nutella setup.
Right now we are building Nutella docker image in this (weather) repositiory.
We want to build that image in the `nutella_framework` repository, and then publish / use
the image from dockerhub.

The Dockerfile build is using untrusted GPG keys that prevent the image from being created on Dockerhub. It should be possible to install Node and GOSU without these keys by chaning the build steps.